Move baseVPNConfiguration to TransientStore
Simplifies configuration update.
